Refactor Constants
