+++ component/src/main/java/org/ballerinalang/net/jms/JMSListenerImpl.java
@@ -22,7 +22,6 @@ import org.ballerinalang.connector.api.ConnectorFuture;
 import org.ballerinalang.connector.api.ConnectorFutureListener;
 import org.ballerinalang.connector.api.Executor;
 import org.ballerinalang.connector.api.Resource;
-import org.ballerinalang.net.jms.actions.utils.Constants;
 import org.wso2.carbon.transport.jms.callback.JMSCallback;
 import org.wso2.carbon.transport.jms.contract.JMSListener;
