+++ component/src/main/java/org/ballerinalang/net/jms/actions/Send.java
@@ -31,9 +31,9 @@ import org.ballerinalang.natives.annotations.Attribute;
 import org.ballerinalang.natives.annotations.BallerinaAction;
 import org.ballerinalang.natives.annotations.BallerinaAnnotation;
 import org.ballerinalang.natives.annotations.ReturnType;
+import org.ballerinalang.net.jms.Constants;
 import org.ballerinalang.net.jms.JMSTransactionContext;
 import org.ballerinalang.net.jms.JMSUtils;
-import org.ballerinalang.net.jms.actions.utils.Constants;
 import org.ballerinalang.util.DistributedTxManagerProvider;
 import org.ballerinalang.util.exceptions.BallerinaException;
 import org.slf4j.Logger;
