+++ component/src/test/java/org/ballerinalang/net/jms/nativeimpl/functions/NetJMSTest.java
@@ -23,8 +23,8 @@ import org.ballerinalang.connector.api.ConnectorFutureListener;
 import org.ballerinalang.connector.impl.BServerConnectorFuture;
 import org.ballerinalang.model.values.BString;
 import org.ballerinalang.model.values.BValue;
+import org.ballerinalang.net.jms.Constants;
 import org.ballerinalang.net.jms.JMSConnectorFutureListener;
-import org.ballerinalang.net.jms.actions.utils.Constants;
 import org.ballerinalang.net.jms.nativeimpl.util.BTestUtils;
 import org.ballerinalang.net.jms.nativeimpl.util.CompileResult;
 import org.ballerinalang.net.jms.nativeimpl.util.TestAcknowledgementCallback;
